move from to picker up one component
